+++ src/game/Object/Unit.cpp
@@ -5518,374 +5518,6 @@ void Unit::SetDeathState(DeathState s)
     m_deathState = s;
 }
 
-/*########################################
-########                          ########
-########       AGGRO SYSTEM       ########
-########                          ########
-########################################*/
-
-bool Unit::CanHaveThreatList(bool ignoreAliveState/*=false*/) const
-{
-    // only creatures can have threat list
-    if (GetTypeId() != TYPEID_UNIT)
-    {
-        return false;
-    }
-
-    // only alive units can have threat list
-    if (!IsAlive() && !ignoreAliveState)
-    {
-        return false;
-    }
-
-    Creature const* creature = ((Creature const*)this);
-
-    // totems can not have threat list
-    if (creature->IsTotem())
-    {
-        return false;
-    }
-
-    // pets can not have a threat list, unless they are controlled by a creature
-    if (creature->IsPet() && creature->GetOwnerGuid().IsPlayer())
-    {
-        return false;
-    }
-
-    // charmed units can not have a threat list if charmed by player
-    if (creature->GetCharmerGuid().IsPlayer())
-    {
-        return false;
-    }
-
-    return true;
-}
-
-//======================================================================
-
-float Unit::ApplyTotalThreatModifier(float threat, SpellSchoolMask schoolMask)
-{
-    if (!HasAuraType(SPELL_AURA_MOD_THREAT))
-    {
-        return threat;
-    }
-
-    if (schoolMask == SPELL_SCHOOL_MASK_NONE)
-    {
-        return threat;
-    }
-
-    SpellSchools school = GetFirstSchoolInMask(schoolMask);
-
-    return threat * m_threatModifier[school];
-}
-
-//======================================================================
-
-void Unit::AddThreat(Unit* pVictim, float threat /*= 0.0f*/, bool crit /*= false*/, SpellSchoolMask schoolMask /*= SPELL_SCHOOL_MASK_NONE*/, SpellEntry const* threatSpell /*= NULL*/)
-{
-    // Only mobs can manage threat lists
-    if (CanHaveThreatList())
-    {
-        m_ThreatManager.addThreat(pVictim, threat, crit, schoolMask, threatSpell);
-    }
-}
-
-//======================================================================
-
-void Unit::DeleteThreatList()
-{
-    if (CanHaveThreatList(true) && !m_ThreatManager.isThreatListEmpty())
-    {
-        SendThreatClear();
-    }
-
-    m_ThreatManager.clearReferences();
-}
-
-//======================================================================
-
-void Unit::TauntApply(Unit* taunter)
-{
-    MANGOS_ASSERT(GetTypeId() == TYPEID_UNIT);
-
-    if (!taunter || (taunter->GetTypeId() == TYPEID_PLAYER && ((Player*)taunter)->isGameMaster()))
-    {
-        return;
-    }
-
-    if (!CanHaveThreatList())
-    {
-        return;
-    }
-
-    Unit* target = getVictim();
-
-    if (target && target == taunter)
-    {
-        return;
-    }
-
-    // Only attack taunter if this is a valid target
-    if (!hasUnitState(UNIT_STAT_STUNNED | UNIT_STAT_DIED) && !IsSecondChoiceTarget(taunter, true))
-    {
-        if (GetTargetGuid() || !target)
-        {
-            SetInFront(taunter);
-        }
-
-        if (((Creature*)this)->AI())
-        {
-            ((Creature*)this)->AI()->AttackStart(taunter);
-        }
-    }
-
-    m_ThreatManager.tauntApply(taunter);
-}
-
-//======================================================================
-
-void Unit::TauntFadeOut(Unit* taunter)
-{
-    MANGOS_ASSERT(GetTypeId() == TYPEID_UNIT);
-
-    if (!taunter || (taunter->GetTypeId() == TYPEID_PLAYER && ((Player*)taunter)->isGameMaster()))
-    {
-        return;
-    }
-
-    if (!CanHaveThreatList())
-    {
-        return;
-    }
-
-    Unit* target = getVictim();
-
-    if (!target || target != taunter)
-    {
-        return;
-    }
-
-    if (m_ThreatManager.isThreatListEmpty())
-    {
-        m_fixateTargetGuid.Clear();
-
-        if (((Creature*)this)->AI())
-        {
-            ((Creature*)this)->AI()->EnterEvadeMode();
-        }
-
-        if (InstanceData* mapInstance = GetInstanceData())
-        {
-            mapInstance->OnCreatureEvade((Creature*)this);
-        }
-
-        if (m_isCreatureLinkingTrigger)
-        {
-            GetMap()->GetCreatureLinkingHolder()->DoCreatureLinkingEvent(LINKING_EVENT_EVADE, (Creature*)this);
-        }
-
-        return;
-    }
-
-    m_ThreatManager.tauntFadeOut(taunter);
-    target = m_ThreatManager.getHostileTarget();
-
-    if (target && target != taunter)
-    {
-        if (GetTargetGuid())
-        {
-            SetInFront(target);
-        }
-
-        if (((Creature*)this)->AI())
-        {
-            ((Creature*)this)->AI()->AttackStart(target);
-        }
-    }
-}
-
-//======================================================================
-/// if pVictim is given, the npc will fixate onto pVictim, if NULL it will remove current fixation
-void Unit::FixateTarget(Unit* pVictim)
-{
-    if (!pVictim)                                           // Remove Fixation
-    {
-        m_fixateTargetGuid.Clear();
-    }
-    else if (pVictim->IsTargetableForAttack())              // Apply Fixation
-    {
-        m_fixateTargetGuid = pVictim->GetObjectGuid();
-    }
-
-    // Start attacking the fixated target or the next proper one
-    SelectHostileTarget();
-}
-
-//======================================================================
-
-bool Unit::IsSecondChoiceTarget(Unit* pTarget, bool checkThreatArea) const
-{
-    MANGOS_ASSERT(pTarget && GetTypeId() == TYPEID_UNIT);
-
-    return
-        pTarget->IsImmunedToDamage(GetMeleeDamageSchoolMask()) ||
-        pTarget->hasNegativeAuraWithInterruptFlag(AURA_INTERRUPT_FLAG_DAMAGE) ||
-        checkThreatArea && ((Creature*)this)->IsOutOfThreatArea(pTarget);
-}
-
-//======================================================================
-
-bool Unit::SelectHostileTarget()
-{
-    // function provides main threat functionality
-    // next-victim-selection algorithm and evade mode are called
-    // threat list sorting etc.
-
-    MANGOS_ASSERT(GetTypeId() == TYPEID_UNIT);
-
-    if (!this->IsAlive())
-    {
-        return false;
-    }
-
-    // This function only useful once AI has been initialized
-    if (!((Creature*)this)->AI())
-    {
-        return false;
-    }
-
-    Unit* target = NULL;
-    Unit* oldTarget = getVictim();
-
-    // first check if we should fixate a target
-    if (m_fixateTargetGuid)
-    {
-        if (oldTarget && oldTarget->GetObjectGuid() == m_fixateTargetGuid)
-        {
-            target = oldTarget;
-        }
-        else
-        {
-            Unit* pFixateTarget = GetMap()->GetUnit(m_fixateTargetGuid);
-            if (pFixateTarget && pFixateTarget->IsAlive() && !IsSecondChoiceTarget(pFixateTarget, true))
-            {
-                target = pFixateTarget;
-            }
-        }
-    }
-    // then checking if we have some taunt on us
-    if (!target)
-    {
-        const AuraList& tauntAuras = GetAurasByType(SPELL_AURA_MOD_TAUNT);
-        Unit* caster;
-
-        // Find first available taunter target
-        // Auras are pushed_back, last caster will be on the end
-        for (AuraList::const_reverse_iterator aura = tauntAuras.rbegin(); aura != tauntAuras.rend(); ++aura)
-        {
-            if ((caster = (*aura)->GetCaster()) && caster->IsInMap(this) &&
-                caster->IsTargetableForAttack() && caster->isInAccessablePlaceFor((Creature*)this) &&
-                !IsSecondChoiceTarget(caster, true))
-            {
-                target = caster;
-                break;
-            }
-        }
-    }
-
-    // No valid fixate target, taunt aura or taunt aura caster is dead, standard target selection
-    if (!target && !m_ThreatManager.isThreatListEmpty())
-    {
-        target = m_ThreatManager.getHostileTarget();
-    }
-
-    if (target)
-    {
-        if (!hasUnitState(UNIT_STAT_STUNNED | UNIT_STAT_DIED))
-        {
-            // PACIFIED creatures (training dummies, etc.) keep their spawn
-            // orientation. PACIFIED already gates attack initiation, so visual
-            // auto-facing here is purely cosmetic AND it falsifies the angle-of-
-            // attack rules in RollMeleeOutcomeAgainst (the from-behind gate for
-            // parry/block) for any player testing combat against the creature.
-            if (!HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PACIFIED))
-            {
-                SetInFront(target);
-            }
-            if (oldTarget != target)
-            {
-                ((Creature*)this)->AI()->AttackStart(target);
-            }
-
-            // check if currently selected target is reachable
-            // NOTE: path alrteady generated from AttackStart()
-            if (!GetMotionMaster()->GetCurrent()->IsReachable())
-            {
-                // remove all taunts
-                RemoveSpellsCausingAura(SPELL_AURA_MOD_TAUNT);
-
-                if (m_ThreatManager.getThreatList().size() < 2)
-                {
-                    // only one target in list, we have to evade after timer
-                    // TODO: make timer - inside Creature class
-                    ((Creature*)this)->AI()->EnterEvadeMode();
-                }
-                else
-                {
-                    // remove unreachable target from our threat list
-                    // next iteration we will select next possible target
-                    m_HostileRefManager.deleteReference(target);
-                    m_ThreatManager.modifyThreatPercent(target, -101);
-
-                    // remove target from current attacker, do not exit combat settings
-                    AttackStop(true);
-                }
-
-                return false;
-            }
-        }
-        return true;
-    }
-
-    // no target but something prevent go to evade mode
-    if (!IsInCombat() || HasAuraType(SPELL_AURA_MOD_TAUNT))
-    {
-        return false;
-    }
-
-    // last case when creature don't must go to evade mode:
-    // it in combat but attacker not make any damage and not enter to aggro radius to have record in threat list
-    // for example at owner command to pet attack some far away creature
-    // Note: creature not have targeted movement generator but have attacker in this case
-    if (GetMotionMaster()->GetCurrentMovementGeneratorType() != CHASE_MOTION_TYPE)
-    {
-        for (AttackerSet::const_iterator itr = m_attackers.begin(); itr != m_attackers.end(); ++itr)
-        {
-            if ((*itr)->IsInMap(this) && (*itr)->IsTargetableForAttack() && (*itr)->isInAccessablePlaceFor((Creature*)this))
-            {
-                return false;
-            }
-        }
-    }
-
-    // enter in evade mode in other case
-    m_fixateTargetGuid.Clear();
-    ((Creature*)this)->AI()->EnterEvadeMode();
-
-    if (InstanceData* mapInstance = GetInstanceData())
-    {
-        mapInstance->OnCreatureEvade((Creature*)this);
-    }
-
-    if (m_isCreatureLinkingTrigger)
-    {
-        GetMap()->GetCreatureLinkingHolder()->DoCreatureLinkingEvent(LINKING_EVENT_EVADE, (Creature*)this);
-    }
-
-    return false;
-}
-
 //======================================================================
 //======================================================================
 //======================================================================
@@ -8714,60 +8346,6 @@ uint32 Unit::GetCombatRatingDamageReduction(CombatRating cr, float rate, float c
         percent = cap;
     }
     return uint32(percent * damage / 100.0f);
-}
-
-void Unit::SendThreatUpdate()
-{
-    ThreatList const& tlist = GetThreatManager().getThreatList();
-    if (uint32 count = tlist.size())
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_COMBAT, "WORLD: Send SMSG_THREAT_UPDATE Message");
-        WorldPacket data(SMSG_THREAT_UPDATE, 8 + count * 8);
-        data << GetPackGUID();
-        data << uint32(count);
-        for (ThreatList::const_iterator itr = tlist.begin(); itr != tlist.end(); ++itr)
-        {
-            data << (*itr)->getUnitGuid().WriteAsPacked();
-            data << uint32((*itr)->getThreat());
-        }
-        SendMessageToSet(&data, false);
-    }
-}
-
-void Unit::SendHighestThreatUpdate(HostileReference* pHostilReference)
-{
-    ThreatList const& tlist = GetThreatManager().getThreatList();
-    if (uint32 count = tlist.size())
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_COMBAT, "WORLD: Send SMSG_HIGHEST_THREAT_UPDATE Message");
-        WorldPacket data(SMSG_HIGHEST_THREAT_UPDATE, 8 + 8 + count * 8);
-        data << GetPackGUID();
-        data << pHostilReference->getUnitGuid().WriteAsPacked();
-        data << uint32(count);
-        for (ThreatList::const_iterator itr = tlist.begin(); itr != tlist.end(); ++itr)
-        {
-            data << (*itr)->getUnitGuid().WriteAsPacked();
-            data << uint32((*itr)->getThreat());
-        }
-        SendMessageToSet(&data, false);
-    }
-}
-
-void Unit::SendThreatClear()
-{
-    DEBUG_FILTER_LOG(LOG_FILTER_COMBAT, "WORLD: Send SMSG_THREAT_CLEAR Message");
-    WorldPacket data(SMSG_THREAT_CLEAR, 8);
-    data << GetPackGUID();
-    SendMessageToSet(&data, false);
-}
-
-void Unit::SendThreatRemove(HostileReference* pHostileReference)
-{
-    DEBUG_FILTER_LOG(LOG_FILTER_COMBAT, "WORLD: Send SMSG_THREAT_REMOVE Message");
-    WorldPacket data(SMSG_THREAT_REMOVE, 8 + 8);
-    data << GetPackGUID();
-    data << pHostileReference->getUnitGuid().WriteAsPacked();
-    SendMessageToSet(&data, false);
 }
 
 struct StopAttackFactionHelper
